requirements.txt
duckdb>=1.0.0
pandas>=2.0.0
numpy>=1.25.0

streamlit>=1.30.0
python-dotenv>=1.0.0

openai>=1.6.0
pydantic>=2.5.0
tenacity>=8.2.0

pytest>=8.0.0
